feat(frontend): separate landing and dashboard routes
